change vat calculation in line_items / adjust product charts

+++ db/seeds.rb
@@ -6,6 +6,15 @@
 #   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
 #   Mayor.create(name: 'Emanuel', city: cities.first)
 
+# Taxes
+remove_taxes = Tax.destroy_all
+create_taxes = Tax.create(
+  [
+    { id: 1, name: "Standardsatz (19%)", value: 0.19 },
+    { id: 2, name: "Mehrwertsteuerbefreit", value: 0 }
+  ]
+)
+
 # ContactRole
 remove_contact_roles = ContactRole.destroy_all
 create_contact_roles = ContactRole.create(
@@ -19,25 +28,27 @@ create_contact_roles = ContactRole.create(
 )
 
 # WorkSteps
-remove_work_steps = WorkStep.destroy_all
-create_work_steps = WorkStep.create(
-  [
-    { id: 1, name: "Zuschnitt Laser" },
-    { id: 2, name: "Zuschnitt Wasserstrahl" },
-    { id: 3, name: "Schaft Vorrichten" },
-    { id: 4, name: "Schaft Vorlegen" },
-    { id: 5, name: "Seitenteil & Vorderblatt" },
-    { id: 6, name: "Ösenteil & Herzchen" },
-    { id: 7, name: "Quartierfutter einnähen & lochen" },
-    { id: 8, name: "Fersenkappe vorformen & einkleben" },
-    { id: 9, name: "Boden einstrobeln & Senkel einziehen" },
-    { id: 10, name: "Mittel- & Außensohle schneiden" },
-    { id: 11, name: "Rohling Anstrich & Verpressen" },
-    { id: 12, name: "Außen- und Mittelsohle rauen" },
-    { id: 13, name: "Schaft & Sohle Anstrich" },
-    { id: 14, name: "Verheiraten" },
-    { id: 15, name: "Endkontrolle", final_step: true },
-    { id: 16, name: "Verpacken" },
-    { id: 17, name: "Zunge (Finale)" }
-  ]
-)
+if WorkStep.count == 0
+  remove_work_steps = WorkStep.destroy_all
+  create_work_steps = WorkStep.create(
+    [
+      { id: 1, name: "Zuschnitt Laser" },
+      { id: 2, name: "Zuschnitt Wasserstrahl" },
+      { id: 3, name: "Schaft Vorrichten" },
+      { id: 4, name: "Schaft Vorlegen" },
+      { id: 5, name: "Seitenteil & Vorderblatt" },
+      { id: 6, name: "Ösenteil & Herzchen" },
+      { id: 7, name: "Quartierfutter einnähen & lochen" },
+      { id: 8, name: "Fersenkappe vorformen & einkleben" },
+      { id: 9, name: "Boden einstrobeln & Senkel einziehen" },
+      { id: 10, name: "Mittel- & Außensohle schneiden" },
+      { id: 11, name: "Rohling Anstrich & Verpressen" },
+      { id: 12, name: "Außen- und Mittelsohle rauen" },
+      { id: 13, name: "Schaft & Sohle Anstrich" },
+      { id: 14, name: "Verheiraten" },
+      { id: 15, name: "Endkontrolle", final_step: true },
+      { id: 16, name: "Verpacken" },
+      { id: 17, name: "Zunge (Finale)" }
+    ]
+  )
+end
